Add server-side auth gates for create pages
